% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-atus.R
\name{date_atus}
\alias{date_atus}
\title{Calculate Date of Accumulated Thermal Units (ATUs)}
\usage{
date_atus(x, atus = 600, start_date = as.Date("1972-03-01"))
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{atus}{A positive number of the accumulated thermal units to exceed.}

\item{start_date}{A Date scalar of the first date
within each year to consider (the year is ignored).
#' If \code{start_date} occurs before the \code{end_date} (when ignoring the year)
then the window is considered  to span two calendar years.}
}
\value{
A tibble with four columns \code{year}, \code{start_date}, \code{end_date} and \code{atus}.
}
\description{
Calculates the date on which a specified number of Accumulated Thermal Units (ATUs)
are exceeded.
}
\examples{
date_atus(gsdd::temperature_data)
}
