% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_days}
\alias{growth_days}
\title{Days Growth Metric}
\usage{
growth_days(vec)
}
\arguments{
\item{vec}{A double vector of mean daily water temperature values.}
}
\value{
A double vector of the days.
}
\description{
Days Growth Metric
}
\examples{
growth_days(c(1,2,3,4,5,6,7,8,9,10))
}
\seealso{
\code{\link[=growth_gdd]{growth_gdd()}}
}
