% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_gdd}
\alias{growth_gdd}
\title{Growing Degree Days Growth Metric}
\usage{
growth_gdd(vec)
}
\arguments{
\item{vec}{A double vector of mean daily water temperature values.}
}
\value{
A double vector of the growing degree days.
}
\description{
Growing Degree Days Growth Metric
}
\examples{
growth_gdd(c(1,2,3,4,5,6,7,8,9,10))
}
\seealso{
\code{\link[=growth_days]{growth_days()}}
}
