% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_pgti_factory}
\alias{growth_pgti_factory}
\title{Generate PGTI Growth Function}
\usage{
growth_pgti_factory(Tmin, Topt, Tmax)
}
\arguments{
\item{Tmin}{A positive real number of the minimum mean daily water temperature for growth.}

\item{Topt}{A positive real number of the optimum mean daily water temperature for growth.}

\item{Tmax}{A positive real number of the maximum mean daily water temperature for growth.}
}
\value{
A double vector of the growth metric.
}
\description{
A function factor to generate PGTI growth functions.
}
\examples{
growth_pgti_factory(3,10,17)(1:20)
}
\seealso{
Other growth:
\code{\link[=growth_days]{growth_days()}},
\code{\link[=growth_degdays]{growth_degdays()}},
\code{\link[=growth_trapezoid_factory]{growth_trapezoid_factory()}}
}
\concept{growth}
