% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_trapezoid_factory}
\alias{growth_trapezoid_factory}
\title{Generate Trapezoid Growth Function}
\usage{
growth_trapezoid_factory(Tmin, Topt, Topt2, Tmax)
}
\arguments{
\item{Tmin}{A positive real number of the minimum mean daily water temperature for growth.}

\item{Topt}{A positive real number of the optimum mean daily water temperature for growth.}

\item{Topt2}{A positive real number of the upper optimum mean daily water temperature for growth.}

\item{Tmax}{A positive real number of the maximum mean daily water temperature for growth.}
}
\value{
A double vector of the growth metric.
}
\description{
A function factor to generate trapezoidal growth functions.
}
\examples{
growth_trapezoid_factory(3,10,12,17)(1:20)
}
\seealso{
Other growth: 
\code{\link{growth_days}()},
\code{\link{growth_gdd}()},
\code{\link{growth_pgti_factory}()}
}
\concept{growth}
