% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd.R
\name{gsdd}
\alias{gsdd}
\title{Calculate Growing Season Degree Days (GSDD)}
\usage{
gsdd(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-12-31"),
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "longest",
  msgs = TRUE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{start_date}{A Date scalar of the first date within each year to consider (the year
is ignored).
#' If \code{start_date} occurs before the \code{end_date} (when ignoring the year) then the window is considered
to span two calendar years.}

\item{end_date}{A Date scalar of the last date within each year to consider (the year is ignored).}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season(s) in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season(s) in C. It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A tibble with two columns \code{year} and \code{gsdd}.
}
\description{
Growing Season Degree Days (GSDD) is a water temperature metric
that is a useful predictor of Cutthroat trout size at the
beginning of winter.
It is the accumulated thermal units (in C)
during the growing season based on the mean daily water temperature values.
}
\details{
The GSDD is calculated across the longest consecutive sequence of non-missing
values which must be at least twice the window width in length otherwise a
missing value is returned.
If the vector includes missing values it is recommended that they are
replaced by estimates of the actual values.

The default values and implementation of the
growing season are based on
Coleman and Fausch (2007) who stated that

We defined the start of the growing season as the
beginning of the first week that average stream temperatures exceeded and
remained above 5C for the season;
the end of the growing season was defined as
the last day of the first week that
average stream temperature dropped below 4C.

For the purposes of the calculation week is assumed to refer to a seven day
rolling average as opposed to the calendar week.

If there are multiple growing 'seasons' within the same year then by
default the returned value is the sum of the GSDD values for \code{"all"} seasons.

The user also has the option to pick the \code{"first"}/\code{"last"} or
\code{"longest"}/\code{"shortest"} season or the season with
the \code{"biggest"}/\code{"smallest"} GSDD.
If the user picks the \code{"longest"} season but there are multiple seasons
with the longest length then the candidate
season with the \code{"biggest"} GSDD is selected.
Conversely in the case of multiple \code{"shortest"} seasons then the
candidate with the \code{"smallest"} GSDD is selected.

Truncation occurs when the start and/or end
of the time series is part way through a growing season.
If the user chooses to ignore truncation then the returned value
very likely to be less than the actual GSDD.
}
\examples{
gsdd(gsdd::temperature_data)
}
\seealso{
\code{\link[=gsdd_vctr]{gsdd_vctr()}}, \code{\link[=gdd]{gdd()}} and \code{\link[=gss]{gss()}}
}
