% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-cf-data.R
\name{gsdd_cf_data}
\alias{gsdd_cf_data}
\title{Calculate Growing Season Degree Days (GSDD) from a Data Frame}
\usage{
gsdd_cf_data(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-12-31"),
  ignore_truncation = FALSE,
  msgs = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.}

\item{start_date}{A Date scalar of the first date within each year to consider (the year
is ignored).}

\item{end_date}{A Date scalar of the last date within each year to consider (the year is ignored).
If \code{end_date} is less than \code{start_date} (ignoring the year) then the window is considered
to span two calendar years.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{msgs}{A flag specifying whether to provide messages.}

\item{...}{Additional arguments passed to \code{\link[=gsdd_cf]{gsdd_cf()}}.}
}
\value{
A tibble with two columns \code{year} and \code{gsdd}.
\code{year}, which is an integer vector, indicates the year in which the window
began and \code{gsdd} which is a non-negative real number provides the GSDD
or a missing value if it cannot be calculated.
}
\description{
The GSDD is calculated for each study year from a data frame
with \code{date} and \code{temperature} columns.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade. For additional information on
GSDD and the various arguments that can be passed via \code{...} see \code{\link[=gsdd_cf]{gsdd_cf()}}.
}
\examples{
data <- gsdd::simulated_data
data$temperature <- data$synthetic
gsdd_cf_data(data)
}
\seealso{
\code{\link[=gsdd_cf]{gsdd_cf()}} and \code{\link[=gdd_cf_data]{gdd_cf_data()}}
}
