% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss.R
\name{gss}
\alias{gss}
\title{Calculate Growing Seasons (GSS)}
\usage{
gss(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-12-31"),
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "longest",
  msgs = TRUE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{start_date}{A Date scalar of the first date within each year to consider (the year
is ignored).
#' If \code{start_date} occurs before the \code{end_date} (when ignoring the year) then the window is considered
to span two calendar years.}

\item{end_date}{A Date scalar of the last date within each year to consider (the year is ignored).}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season(s) in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season(s) in C. It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A tibble with four columns \code{year}, \code{start_dayte}, \code{end_dayte} and \code{gsdd}.
}
\description{
Calculate the number of growing degree days to the specified end date.
}
\examples{
gss(gsdd::temperature_data)
}
\seealso{
\code{\link[=gsdd]{gsdd()}}
}
