% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss-cf-data.R
\name{gss_cf_data}
\alias{gss_cf_data}
\title{Calculate Growing Seasons from a Data Frame}
\usage{
gss_cf_data(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-12-31"),
  ignore_truncation = FALSE,
  msgs = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector of the
mean daily water temperature values for the period
of interest in C. It must be consist of no more than
366 values.}

\item{start_date}{A Date scalar of the first date within each year to consider (the year
is ignored).}

\item{end_date}{A Date scalar of the last date within each year to consider (the year is ignored).
If \code{end_date} is less than \code{start_date} (ignoring the year) then the window is considered
to span two calendar years.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{msgs}{A flag specifying whether to provide messages.}

\item{...}{Additional arguments passed to \code{\link[=gsdd]{gsdd()}}.}
}
\value{
A tibble with four columns \code{year}, \code{start_dayte}, \code{end_dayte} and \code{gsdd}.
}
\description{
Calculate Growing Seasons from a Data Frame
}
\examples{
data <- gsdd::simulated_data
data$temperature <- data$synthetic
gss_cf_data(data)
}
\seealso{
\code{\link[=gsdd_data]{gsdd_data()}}
}
