% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss-plot.R
\name{gss_plot}
\alias{gss_plot}
\title{Plot Growing Seasons (GSS)}
\usage{
gss_plot(
  x,
  start_date = as.Date("1972-03-01"),
  end_date = as.Date("1972-11-30"),
  min_length = NULL,
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "all",
  nrow = NULL,
  ncol = NULL,
  msgs = TRUE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{start_date}{A Date scalar of the first date
within each year to consider (the year is ignored).
#' If \code{start_date} occurs before the \code{end_date} (when ignoring the year)
then the window is considered  to span two calendar years.}

\item{end_date}{A Date scalar of the last date
within each year to consider (the year is ignored).}

\item{min_length}{A whole number of the minimum length to consider.
Must be at least twice the window width but no more than 365.
If NULL then set to be the difference between \code{start_date} and \code{end_date}
(ignoring the year).}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season(s) in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season(s) in C.
It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{nrow}{A count of the number of rows to facet by.}

\item{ncol}{A count of the number of columns to facet by.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A ggplot object.
}
\description{
Plots the growing seasons in a year.
}
\examples{
\dontrun {
  gss_plot(gsdd::temperature_data)
}
}
\seealso{
\code{\link[=gss]{gss()}}
}
