% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for functions}
\arguments{
\item{complete}{A flag specifying whether the vector of water temperatures
represents the complete possible growing period (by default FALSE).
If TRUE a growing season is not considered to be truncated
at the start and/or end if the water temperature is above the threshold
at the start and/or ends.}

\item{end_date}{A Date scalar of the last date
within each year to consider (the year is ignored).}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season(s) in C.
It must be greater than or equal to
the start temperature.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{min_length}{A whole number of the minimum length to consider.
Must be at least twice the window width but no more than 365.
If NULL then set to be the difference between \code{start_date} and \code{end_date}
(ignoring the year).}

\item{msgs}{A flag specifying whether to provide messages.}

\item{ncol}{A count of the number of columns to facet by.}

\item{nrow}{A count of the number of rows to facet by.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{start_date}{A Date scalar of the first date
within each year to consider (the year is ignored).
#' If \code{start_date} occurs before the \code{end_date} (when ignoring the year)
then the window is considered  to span two calendar years.}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season(s) in C.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}
}
\description{
Descriptions of the parameters for functions
}
\keyword{internal}
