% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Sample Data for GSDD Calculation Demonstration}
\format{
A tibble with columns:
\describe{
\item{Date}{date for each calendar year}
\item{synthetic}{synthetic yearly temperature data for demonstration and
testing}
}
}
\usage{
simulated_data
}
\description{
The data is a synthetic time series of daily temperature data for the year
2019.
}
\keyword{datasets}
