% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdd.R
\name{gdd}
\alias{gdd}
\title{Calculate Growing Degree Days (GDD) from a Data Frame}
\usage{
gdd(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-09-30"),
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "longest",
  msgs = TRUE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.}

\item{start_date}{A Date scalar of the first date within each year to consider (the year
is ignored).}

\item{end_date}{A Date scalar of the end date for the growing degree days
(the year is ignored).
If \code{end_date} is less than \code{start_date} (ignoring the year) then the window is considered
to span two calendar years.}

\item{ignore_truncation}{A flag specifying whether to ignore start truncation (end truncation is always ignored).}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season(s) in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season(s) in C. It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A tibble with two columns \code{year} and \code{gdd}.
\code{year}, which is an integer vector, indicates the year in which the window
began and \code{gdd} which is a non-negative real number provides the GSDD
or a missing value if it cannot be calculated.
}
\description{
The GDD which is a the growing degree days to the end date
is calculated for each study year from a data frame.
with \code{date} and \code{temperature} columns.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.
}
\examples{
data <- gsdd::simulated_data
data$temperature <- data$synthetic
gdd(data)
}
\seealso{
\code{\link[=gsdd]{gsdd()}}
}
