% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_days}
\alias{growth_days}
\title{Days Growth Metric}
\usage{
growth_days(vec, Tmin = 0)
}
\arguments{
\item{vec}{A numeric vector of mean daily water temperature values.}

\item{Tmin}{A positive real number of the minimum mean daily water temperature for growth.}
}
\value{
A double vector of the growth metric.
}
\description{
Days Growth Metric
}
\examples{
growth_days(c(1,2,3,4,5,6,7,8,9,10))
}
\seealso{
Other growth: 
\code{\link{growth_degdays}()},
\code{\link{growth_pgti_factory}()},
\code{\link{growth_trapezoid_factory}()}
}
\concept{growth}
