% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_days}
\alias{growth_days}
\title{Days Growth Metric}
\usage{
growth_days(vec)
}
\arguments{
\item{vec}{A numeric vector of mean daily water temperature values.}
}
\value{
A double vector of the growth metric.
}
\description{
Days with a negative mean daily water temperature are assigned 0 growth.
}
\examples{
growth_days(c(1,2,3,4,5,6,7,8,9,10))
}
\seealso{
Other growth: 
\code{\link{growth_gdd}()},
\code{\link{growth_pgti_factory}()},
\code{\link{growth_trapezoid_factory}()}
}
\concept{growth}
