% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-package.R
\docType{package}
\name{gsdd-package}
\alias{gsdd-package}
\title{gsdd: Calculate Growing Season Degree Days from Water Temperature Data}
\description{
Calculates Growing Season Degree Days (GSDD), Growing Degree Days (GDD) and Growing Seasons (GSS) from water temperature data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/gsdd}
  \item Report bugs at \url{https://github.com/poissonconsulting/gsdd/issues/}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Sarah Lyons (\href{https://orcid.org/0000-0002-6745-6796}{ORCID})
}

Other contributors:
\itemize{
  \item Jill Brooks [contributor]
  \item Bronwen Lewis [contributor]
  \item Teck Resources [funder, copyright holder]
}

}
\keyword{internal}
