% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-vctr.R
\name{gsdd_vctr}
\alias{gsdd_vctr}
\title{Calculate Growing Season Degree Days (GSDD) from a vector}
\usage{
gsdd_vctr(
  x,
  ignore_truncation = FALSE,
  min_length = 274,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "all",
  msgs = TRUE
)
}
\arguments{
\item{x}{A numeric vector of the
mean daily water temperature values for the period
of interest in C.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{min_length}{A whole number of the minimum length to consider.
Must be at least twice the window width but no more than 365.
If NULL then set to be the difference between \code{start_date} and \code{end_date}
(ignoring the year).}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season(s) in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season(s) in C.
It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A non-negative real number of the GSDD.
}
\description{
Calculate Growing Season Degree Days (GSDD) from a vector
}
\examples{
gsdd_vctr(c(rep(1, 10), rep(10, 20), rep(1, 200)))
gsdd_vctr(gsdd::temperature_data$temperature)
}
\seealso{
\code{\link[=gsdd]{gsdd()}}
}
