% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_data}
\alias{temperature_data}
\title{Mean Daily Water Temperature Data}
\format{
A tibble with columns:
\describe{
\item{date}{A Date vector.}
\item{temperature}{A real vector of mean daily water temperature values.}
}
}
\usage{
temperature_data
}
\description{
Mean Daily Water Temperature Data
}
\keyword{datasets}
